guest-contents: correctly update frame sink id registration
This change fixes 2 issues with updating frame sink id registration when
attaching or detaching inner WebContents.

The first is that RenderWidgetHostViewChildFrame::UnregisterFrameSinkId
which RecursivelyUnregisterRenderWidgetHostViews calls to unregister
frame sink id from input event router no longer unregister frame sink id.
This probably is a regression from http://crrev.com/c/4912959 where we
now simply resync registration with ownership. As ownership doesn't
change, unregistration is not done.
To fix the issue, revert code back to before http://crrev.com/c/4912959
which unregisters frame sink id without checking ownership.

The second issue is that RecursivelyRegisterRenderWidgetHostViews and
RecursivelyUnregisterRenderWidgetHostViews does not handles views for
pending navigation when it tries to find all views in the inner
WebContents. When RenderWidgetHostViewChildFrame is created for the
speculative render frame host during navigation, it is set as frame
sink id owner when RenderWidgetHostViewChildFrame::Init calls
RenderWidgetHostImpl::SetView. After it is set as owner, it will not
try to update sink frame id registration later on when navigation is
committed as there is no ownership change.
To fix the issue, we update GetRenderWidgetHostViewsInWebContentsTree
to include speculative render frame hosts. Some of the views for the
speculative render frame hosts might not be a owner. For those views the
code will naturally be a no op.

Bug: 393144165
Change-Id: I6894d81a818e494b4e7330aa22eb952f5d67cea9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6942986
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Liang Zhao <lzhao@microsoft.com>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1517455}