Fix components_unittests component build
The previous change to make guest_contents a component meant that the
component build of components_unittests was linking against the :browser
component DLL instead of the original source. As only symbols used by
the :browser's sources are included in the component DLL, other guest
contents symbols were missing. Adding an explicit deps for the
unit_tests fixes the issue.