Add SecureEmbed::SetSurfaceId support
This CL adds SurfaceId embedding support to SecureEmbedWebPlugin via
cc::SurfaceLayer. Details follow:

Mojom updates:
- Add SetSurfaceId(viz::SurfaceId) method to SecureEmbed.

Renderer side (SecureEmbedWebPlugin):
- Implement SetSurfaceId() with cc::SurfaceLayer for compositor
  embedding. The first call will disable placeholder painting.

Future changes are expected to provide a surfaceId from the
SecureEmbedHost which identifies the surface of the attached
WebContents.

This change also adds a test to verify the above functionality. The
test creates CompositorFrames that draw a colored rectangle. The test
attaches those frames, via their surface IDs, and verifies the resulting
rendered output for the embed element is the appropriate color. Once
there's support for attaching a WebContents' surface ID, these tests can
be replaced with drawing verification of the attached WebContents.