Fix DSF and optimize Send/SynchronizeVisualProperties
Previously, SendVisualProperties would always send the properties to the
embedded WebContents, even if nothing changed. This was partially fixed
in #49 for embedded content's local surface ID updates. This change
expands on that to avoid unnecessary updates for a much larger set of
properties, similar to RemoteFrame. As part of that it also renames
SendVisualProperties to SynchronizeVisualProperties.

Included in this PR is a modification to how device scale factor is
calculated, so that it's based on the ancestor widget (top-most local
frame). Local css zoom on the embed element isn't applied, but dragging
the window between two different desktops with different device scale
factors is working now.