Use GlobalRenderFrameHostId instead of RFH raw_ptr
Replace raw_ptr<RenderFrameHost> with GlobalRenderFrameHostId in
SecureEmbedHost to avoid storing a raw pointer to a RenderFrameHost.
This avoids dangling pointer issues when the RFH is destroyed down
before the mojom connection teardown from the plugin causes the
SecureEmbedHost to be destroyed.

After this change, the browsertests can be ran without failing due to
dangling ptrs.