RFC: Hacky/incorrect way of making BG tab capture work.
The obvious incorrect part here is the change to
RenderWidgetHostViewChildFrame::ShowWithVisibility;
CanBecomeVisible returns false because:

frame_connector_->IsHidden()

returns true; there needs to be some thinking of what
exactly these methods are supposed to mean to come up
with the proper approach.

(This path is invoked via
 content::WebContentsImpl::IncrementCapturerCount ->
 content::WebContentsImpl::UpdateVisibilityAndNotifyPageAndView).

We also seem to renderer a phantom cursor for the BG frame,
which is wrong, but I have no clue on since it took me forever
for stuff to show up.